LayoutTests/inspector/styles-disable-then-delete.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="elements-tests.js"></script>
<script>

function doit()
{
    function callback(result)
    {
        dumpArray(result);
        notifyDone();
    }
    evaluateInWebInspector("frontend_runTest", callback);
}

// Frontend functions.

function frontend_runTest(testController)
{
    testController.waitUntilDone();
    frontend_expandDOMSubtreeAndRun(testController, null, frontend_selectContainerElementContinuation);
}

function frontend_selectContainerElementContinuation(testController)
{
    frontend_selectElementAndRun(testController, "container", frontend_disableContainerStyle);
}

function frontend_disableContainerStyle(testController)
{
    var treeItem = frontend_getElementStylePropertyTreeItem("font-weight");
    testController.results.push("Before disable");
    testController.results = testController.results.concat(frontend_getSelectedElementStyles(true, true));

    treeItem.toggleEnabled({ target: { checked: false } });
    testController.runAfterPendingDispatches(frontend_deleteContainerStyle.bind(this, testController));
}

function frontend_deleteContainerStyle(testController)
{
    var treeItem = frontend_getElementStylePropertyTreeItem("font-weight");
    testController.results.push("After disable");
    testController.results = testController.results.concat(frontend_getSelectedElementStyles(true, true));

    treeItem.applyStyleText("", true);
    testController.runAfterPendingDispatches(frontend_reselectNodeAndDumpResults.bind(this, testController));
}

function frontend_reselectNodeAndDumpResults(testController)
{
    function continuation()
    {
        frontend_selectElementAndRun(testController, "container", frontend_dumpResults);
    }
    frontend_selectElementAndRun(testController, "other", continuation);
}

function frontend_dumpResults(testController)
{
    var treeItem = frontend_getElementStylePropertyTreeItem("font-weight");
    testController.results.push("After delete");
    testController.results = testController.results.concat(frontend_getSelectedElementStyles(true, true));
    testController.notifyDone();
}
</script>
</head>

<body onload="onload()">
<p>
Tests that removal of property following its disabling works.
</p>

<div id="container" style="font-weight:bold">
</div>

<div id="other">
</div>

</body>
</html>
