LayoutTests/inspector/styles-disable-then-delete-expected.txt

Tests that removal of property following its disabling works.

Before disable
[expanded] element.style { (element’s “style” attribute)
font-weight: bold;


After disable
[expanded] element.style { (element’s “style” attribute)
/-- overloaded --/ /-- disabled --/ font-weight: bold;


After delete
[expanded] element.style { (element’s “style” attribute)
